adding btree_map to benchmark and other changes

+++ src/libpmemobj/tx.c
@@ -159,7 +159,7 @@ long long get_hard_treshold(long long val){
 	double cntr = 0;
 
 	cntr = (double)val;
-	cntr =  (cntr * (EAP_BUDGET_THRESHOLD/100));
+	cntr =  (cntr * (double)((double)EAP_BUDGET_THRESHOLD/100));
 	printf("cntr %lf\n",cntr);
 	return (long long)cntr;
 }
